ThinkHDAPS: Changed how disks that can unload are detected
Previously iterated through a list of possible disks and checked if its entry
existed in /sys/block/. Now it checks for every entry in /sys/block whether
a queue/protect_method file exists and that it contains "[unload]". Entries
that meet these conditions should be able to unload. Currently it takes the
first disk that is found capable.