Remove vala-test applet
It appears that vala integration does not fully support out-of-source
builds: it breaks the 'dist' target if the specific subdir is not added
to the build list, which is the case for vala-test.

vala support will be deprecated anyway, thus no need to keep broken
applet which is not useful for the users anyway.