Fix cairo-menu.
Cairo-menu had several segmentation faults that did not let it start.
Mainly due to the fact that gnomecc.menu no longer exists in gnome3.
The code for this was removed.
.
Grey out the sessions submenu if no Logout/Lockout/Shutdown program was found.
.
Replaced the deprecated show_all to show for an easier migration to gtk3.
.
Fixed the session menu for gnome, cinnamon and mate. Gnome3 requires
activating the gnome-screensaver service (possibly manually since it's
deprecated). Cinnamon needs cinnamon-screensaver and possibly starting
AWN with a 10-15 sec delay since the screensaver service starts delayed.
.
Remove the "if (!main_menu_tree)" since everytime a new program got installed
or removed the act of adding/removing the .desktop file would crash the menu.